spanify: Handle `subspan()` from `operator+=()`
Coerce `subspan()` arguments as necessary when emitting these for
expressions like `a += b`.

Fixed: 412395846
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6640537
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1476004}

+++ tools/clang/spanify/tests/assignment-expected.cc
@@ -3,11 +3,13 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 #include <array>
+#include <cstdint>
 #include <vector>
 
 #include "base/containers/span.h"
 #include "base/memory/raw_ptr.h"
 #include "base/memory/raw_span.h"
+#include "base/numerics/safe_conversions.h"
 
 // Expected rewrite:
 // base::span<T> get()
@@ -80,7 +82,13 @@ void fct() {
   // gg = (condition) ? ctn1 : ctn2;
   gg = (condition) ? ctn1 : ctn2;
 
-  gg = gg.subspan(1);  // Buffer usage, leads gg to be rewritten.
+  // Buffer usage, leads gg to be rewritten.
+  // Expected rewrite:
+  // gg = gg.subspan(1u);
+  gg = gg.subspan(1u);
 
-  gg = gg.subspan(index * 2);  // Buffer usage
+  // Buffer usage
+  // Expected rewrite:
+  // gg = gg.subspan(base::checked_cast<size_t>(index * 2));
+  gg = gg.subspan(base::checked_cast<size_t>(index * 2));
 }
